requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.28.0
pytest>=7.0.0
python-dotenv>=1.0.0
aws-cdk.aws-amplify-alpha